Add hook to check link import
